dataScience/ingestaMongo/requirements.txt
pymongo
faker
pandas
psycopg2-binary
boto3
python-dotenv